Set permissions with one chown command
This went from 40 seconds to 0.5 seconds in my one test.
The previous 5 find commands found every file that was not owned by APACHE_USER and then started a a chown command to fix those individual files. In the end, everything is owned by APACHE_USER, and running "chown -R" does the same thing in one process.